requirements.txt
groq
tiktoken 
chromadb
langchain
langchain-core
langchain-community
flask
GitPython
python-dotenv 
sentence-transformers
langchain_groq
-e .